+++ UnityProject/Trecs/Assets/com.trecs.core/Scripts/Heap/FrameScopedNativeSharedHeap.cs
@@ -266,7 +266,7 @@ namespace Trecs
             _idCounter.EnsureAtLeast(reader.Read<uint>("IdCounter"));
             var numEntries = reader.Read<int>("NumEntries");
 
-            _entries.EnsureCapacity((uint)numEntries);
+            _entries.EnsureCapacity(numEntries);
 
             uint maxAddress = 0;
